Feat/handle expire for new ongoing application (#377)
* fix(constants): new constant for ongoing application

* fix(dynamodb): updateCaseExpirationTime to dynamodb module

* fix(caseexpirytime): to its own module

* fix(syncexpirytime): to typescript and use the new modules

* fix(tsconfig.json): add newline

* fix(constants.d.ts): removal of duplicate constant
