+++ services/viva-ms/src/helpers/caseExpiryTime.ts
@@ -1,0 +1,60 @@
+import {
+  ACTIVE_ONGOING,
+  ACTIVE_ONGOING_NEW_APPLICATION,
+  ACTIVE_SIGNATURE_PENDING,
+  ACTIVE_SIGNATURE_COMPLETED,
+  ACTIVE_SUBMITTED,
+  ACTIVE_PROCESSING,
+  ACTIVE_COMPLETION_ONGOING,
+  ACTIVE_RANDOM_CHECK_ONGOING,
+  ACTIVE_COMPLETION_SUBMITTED,
+  ACTIVE_RANDOM_CHECK_SUBMITTED,
+  ACTIVE_COMPLETION_REQUIRED_VIVA,
+  ACTIVE_RANDOM_CHECK_REQUIRED_VIVA,
+  CLOSED_APPROVED_VIVA,
+  CLOSED_PARTIALLY_APPROVED_VIVA,
+  CLOSED_REJECTED_VIVA,
+  CLOSED_COMPLETION_REJECTED_VIVA,
+  SIX_MONTHS_IN_HOURS,
+  FORTY_FIVE_DAYS_IN_HOURS,
+  SEVENTY_TWO_HOURS,
+} from '../libs/constants';
+
+const statusHourMap: Record<string, number> = {
+  [ACTIVE_ONGOING]: SEVENTY_TWO_HOURS,
+  [ACTIVE_ONGOING_NEW_APPLICATION]: SEVENTY_TWO_HOURS,
+
+  [ACTIVE_SIGNATURE_PENDING]: FORTY_FIVE_DAYS_IN_HOURS,
+  [ACTIVE_SIGNATURE_COMPLETED]: FORTY_FIVE_DAYS_IN_HOURS,
+
+  [ACTIVE_SUBMITTED]: FORTY_FIVE_DAYS_IN_HOURS,
+  [ACTIVE_PROCESSING]: FORTY_FIVE_DAYS_IN_HOURS,
+
+  [ACTIVE_COMPLETION_SUBMITTED]: FORTY_FIVE_DAYS_IN_HOURS,
+  [ACTIVE_RANDOM_CHECK_SUBMITTED]: FORTY_FIVE_DAYS_IN_HOURS,
+
+  [ACTIVE_COMPLETION_ONGOING]: FORTY_FIVE_DAYS_IN_HOURS,
+  [ACTIVE_RANDOM_CHECK_ONGOING]: FORTY_FIVE_DAYS_IN_HOURS,
+
+  [ACTIVE_COMPLETION_REQUIRED_VIVA]: FORTY_FIVE_DAYS_IN_HOURS,
+  [ACTIVE_RANDOM_CHECK_REQUIRED_VIVA]: FORTY_FIVE_DAYS_IN_HOURS,
+
+  [CLOSED_APPROVED_VIVA]: SIX_MONTHS_IN_HOURS,
+  [CLOSED_PARTIALLY_APPROVED_VIVA]: SIX_MONTHS_IN_HOURS,
+  [CLOSED_REJECTED_VIVA]: SIX_MONTHS_IN_HOURS,
+  [CLOSED_COMPLETION_REJECTED_VIVA]: SIX_MONTHS_IN_HOURS,
+};
+
+function getHoursOnStatusType(statusType: string): number {
+  const hours = statusHourMap[statusType];
+
+  if (!hours) {
+    throw `Expiry time not set for status: ${statusType}`;
+  }
+
+  return hours;
+}
+
+export default {
+  getHoursOnStatusType,
+};
